Abortable subprocesses.
Until now, when celery task spawned subprocess, it has been imposible to stop them. Eaven when the parent process (the celery process) was SIGKILLed, the child process forked to init. Now, BPMTask inherits from AbortableTask and listens with BPMTask.track_task if the process is aborted.

When Task.revoke is called, first the subprocess is aborted and then the celery process is revoked.